Feat: Seperate OrderStickyNavBar into component
